Separate middleware from pipeline.
